create separate file for product page styled components
